+++ core/modules/modScrumProject.class.php
@@ -65,7 +65,7 @@ class modScrumProject extends DolibarrModules
 		$this->editor_url = 'www.atm-consulting.fr';
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
 
-		$this->version = '1.14.0';
+		$this->version = '1.14.1';
 
 		// Url to the file with your last numberversion of this module
 		//$this->url_last_version = 'http://www.example.com/versionmodule.txt';
@@ -1078,6 +1078,12 @@ class modScrumProject extends DolibarrModules
 	{
 		global $conf, $langs;
 
+		if(intval(DOL_VERSION) < 17){
+			$this->db->query('ALTER TABLE llx_element_element MODIFY COLUMN sourcetype VARCHAR(64) NOT NULL;');
+			$this->db->query('ALTER TABLE llx_element_element MODIFY COLUMN targettype VARCHAR(64) NOT NULL;');
+			$this->db->query('ALTER TABLE llx_c_type_contact MODIFY COLUMN element VARCHAR(64) NOT NULL;');
+		}
+
 		$result = $this->_load_tables('/scrumproject/sql/');
 		if ($result < 0) return -1; // Do not activate module if error 'not allowed' returned when loading module SQL queries (the _load_table run sql with run_sql with the error allowed parameter set to 'default')
 
@@ -1105,11 +1111,6 @@ class modScrumProject extends DolibarrModules
 		$this->remove($options);
 
 
-		if(intval(DOL_VERSION) < 17){
-			$this->db->query('ALTER TABLE llx_element_element MODIFY COLUMN sourcetype VARCHAR(64) NOT NULL;');
-			$this->db->query('ALTER TABLE llx_element_element MODIFY COLUMN targettype VARCHAR(64) NOT NULL;');
-			$this->db->query('ALTER TABLE llx_c_type_contact MODIFY COLUMN element VARCHAR(64) NOT NULL;');
-		}
 
 		$sql = array();